requirements.txt
attrs==23.1.0
backoff==2.2.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
fiona==1.9.5
Flask==3.0.0
Flask-Cors==4.0.0
geopandas==0.14.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
numpy==1.26.2
opencage==2.4.0
packaging==23.2
pandas==2.1.4
protobuf==4.21.12
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
shapely==2.0.2
six==1.16.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
